Add DNA fingerprint
