+++ bangkok/39-GenomTex/templates/record.html
@@ -9,6 +9,7 @@
       <th>File name</th>
       <th>Mutations</th>
       <th>Dist to REF</th>
+      <th>Fingerprint</th>
       <th>Sequenced At</th>
       <th>UUID</th>
       <th>CID</th>
@@ -25,6 +26,9 @@
       </td>
       <td>
           <p class="fw-normal mb-1">{{ record["cos_dist"] }}</p>
+      </td>
+      <td>
+          <p class="fw-normal mb-1">{{ record["dna_fingerprint"] }}</p>
       </td>
       <td>
           <p class="fw-normal mb-1">{{ record["sequenced_at"] }}</p>
